Cast to correct sign

+++ source/core_mqtt.c
@@ -950,7 +950,7 @@ static MQTTStatus_t sendPublishAcks( MQTTContext_t * pContext,
             LogError( ( "Failed to send ACK packet: PacketType=%02x, "
                         "SentBytes=%d, "
                         "PacketSize=%lu.",
-                        ( int ) packetTypeByte,
+                        ( unsigned int ) packetTypeByte,
                         ( int ) bytesSent,
                         MQTT_PUBLISH_ACK_PACKET_SIZE ) );
             status = MQTTSendFailed;
@@ -1198,7 +1198,8 @@ static MQTTStatus_t handleIncomingAck( MQTTContext_t * pContext,
 
     appCallback = pContext->appCallback;
 
-    LogDebug( ( "Received packet of type %02x.", ( int ) pIncomingPacket->type ) );
+    LogDebug( ( "Received packet of type %02x.",
+                ( unsigned int ) pIncomingPacket->type ) );
 
     switch( pIncomingPacket->type )
     {
@@ -1233,7 +1234,7 @@ static MQTTStatus_t handleIncomingAck( MQTTContext_t * pContext,
         default:
             /* Bad response from the server. */
             LogError( ( "Unexpected packet type from server: PacketType=%02x.",
-                        ( int ) pIncomingPacket->type ) );
+                        ( unsigned int ) pIncomingPacket->type ) );
             status = MQTTBadResponse;
             break;
     }
@@ -1501,8 +1502,8 @@ static MQTTStatus_t receiveConnack( const MQTTContext_t * pContext,
         {
             LogError( ( "Incorrect packet type %X received while expecting"
                         " CONNACK(%X).",
-                        ( int ) pIncomingPacket->type,
-                        ( int ) MQTT_PACKET_TYPE_CONNACK ) );
+                        ( unsigned int ) pIncomingPacket->type,
+                        MQTT_PACKET_TYPE_CONNACK ) );
             status = MQTTBadResponse;
         }
     }
